- исправлены нейминги - разложены DTO - поправлен NetworkService
