+++ MVP-UI/MVP-UI/DTO/ComponentsImageDTO.swift
@@ -1,0 +1,10 @@
+// ComponentsImageDTO.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Структура данных ссылки на картинку
+struct ComponentsImageDTO: Codable {
+    /// Ссылка на картинку
+    let url: String
+}
