+++ MVP-UI/MVP-UI/DTO/DetailDTO.swift
@@ -1,0 +1,22 @@
+// DetailDTO.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Детальная струкрутра данных рецептов
+struct DetailDTO: Codable {
+    /// Заголовок
+    let label: String
+    /// Время приготовления
+    let totalTime: Int
+    /// Кол-во калорий
+    let calories: Double
+    /// Картинка рецепта
+    let images: ImagesDTO
+    /// Масса рецепта
+    let totalWeight: Double
+    /// КБЖУ
+    let totalNutrients: [String: TotalDTO]
+    /// Описание рецепта
+    let ingredientLines: [String]
+}
