+++ MVP-UI/MVP-UI/DTO/DetailsResponseDTO.swift
@@ -1,0 +1,10 @@
+// DetailsResponseDTO.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// DetailsResponseDTO
+struct DetailsResponseDTO: Codable {
+    /// Рецепт
+    let recipe: DetailDTO
+}
