+++ MVP-UI/MVP-UI/DTO/RecipeDetailDTO.swift
@@ -1,0 +1,10 @@
+// RecipeDetailDTO.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Детальные рецепты
+struct RecipeDetailDTO: Codable {
+    /// Массив для запрса
+    let hits: [DetailsResponseDTO]
+}
