+++ MVP-UI/MVP-UI/DTO/RecipeHitDTO.swift
@@ -1,0 +1,10 @@
+// RecipeHitDTO.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// RecipeHitDTO
+struct RecipeHitDTO: Codable {
+    /// Конкретный рeцепт
+    let recipe: RecipeDTO
+}
