+++ MVP-UI/MVP-UI/DTO/RecipeResponseDTO.swift
@@ -1,0 +1,10 @@
+// RecipeResponseDTO.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// РецептDTO
+struct RecipeResponseDTO: Codable {
+    /// Массив рецептов
+    let hits: [RecipeHitDTO]
+}
