+++ MVP-UI/MVP-UI/ProfileModule/View/TermsPrivatePolicyView.swift
@@ -40,11 +40,11 @@ final class TermsPrivatePolicyView: UIView {
             "Thank you for being a part of our culinary community! Enjoy exploring and cooking up a storm!"
     }
 
-    /// состояния View
+    /// Состояния View
     enum StateView {
-        /// открыта
+        /// Открыта
         case openView
-        /// закрыта
+        /// Закрыта
         case closeView
     }
